separate css and add a greeting mesage
