+++ orange/src/components/Necessity/NecessityItem/NecessityItem.tsx
@@ -1,16 +1,13 @@
-import React, { Dispatch } from 'react';
-import { MdDelete } from 'react-icons/md';
-import { connect } from 'react-redux';
+import React from 'react';
 import styled from 'styled-components';
-import { necessityActions } from '../../../store/actions';
-import './NecessityItem.css';
+import { MdDelete } from 'react-icons/md';
 
-const Remove = styled.div`
+const NecessityItemRemove = styled.div`
     display: flex;
     align-items: center;
     justify-content: center;
     color: #F1948A;
-    font-size: 21px;
+    font-size: 20px;
     cursor: pointer;
     &:hover {
         color: #B03A2E;
@@ -24,71 +21,39 @@ const NecessityItemBlock = styled.div`
     padding-top: 12px;
     padding-bottom: 12px;
     &:hover {
-        font-weight: bold;
-        cursor: pointer;
-        ${Remove}{
+        ${NecessityItemRemove}{
             display: initial;
         }
     }
 `;
 
-interface Props {
-  cursor?: any;
-  done?: boolean;
-  id: number;
-  name: string;
-  option: string;
-  price: number;
-  remove(id: number): any;
+const Text = styled.div`
+    flex: 1;
+    font-size: 21px;
+    color: #495057;
+    .option {
+      color: #868e96;
+      font-size: 12px;
+    }
 }
+`;
 
-function NecessityItem(props: Props) {
+function NecessityItem({
+  id, name, option, price,
+}: { id: number; name: string; option: string; price: number }) {
   return (
-    <NecessityItemBlock id={`necessity-item-${props.id}`}>
-      {/* 이 자리에 Counter 넣을 것! */}
-      <div className="necessity-text">
-        {`${props.name}`}
-        <span className="necessity-option">
-          {` ${props.option} / ${props.price}원 `}
+    <NecessityItemBlock id={`necessity-item-${id}`}>
+      <Text>
+        {`${name}`}
+        <span className="option">
+          {` ${option} / ${price}원 `}
         </span>
-      </div>
-      <Remove onClick={() => props.remove(props.id)}>
+      </Text>
+      <NecessityItemRemove>
         <MdDelete />
-      </Remove>
+      </NecessityItemRemove>
     </NecessityItemBlock>
-
-  // <div
-  //   className="necessity-item-block"
-  //   id={`necessity-item-${props.id}`}
-  // >
-  //   {/* 이 자리에 Counter 넣을 것! */}
-  //   <div className="necessity-text">
-  //     {`${props.name}`}
-  //     <span className="necessity-option">
-  //       {` ${props.option} / ${props.price}원 `}
-  //     </span>
-  //   </div>
-  //   <button
-  //     className="necessity-remove-button"
-  //     style={{ background: 'none', border: 'none' }}
-  //     type="submit"
-  //     onClick={() => props.remove()}
-  //   >
-  //     <MdDelete />
-  //   </button>
-  // </div>
   );
 }
 
-const mapDispatchToProps = (dispatch: Dispatch<any>) => ({
-  remove: (id: number): void => dispatch(
-    necessityActions.removeNecessity(id),
-  ),
-});
-
-const mapStateToProps = (state: any) => ({
-  removeStatus: state.necessity.removeStatus,
-  me: state.user.me,
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(NecessityItem);
+export default React.memo(NecessityItem);
