+++ orange/src/components/Necessity/NecessityList/NecessityList.css
@@ -1,0 +1,6 @@
+.necessity-list-block {
+    flex: 1;
+    padding: 20px 32px;
+    padding-bottom: 48px;
+    overflow-y: auto;
+  }
