+++ orange/src/components/Necessity/NecessityList/NecessityList.tsx
@@ -1,14 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import styled from 'styled-components';
-import NecessityItem from './NecessityItem';
-
-const NecessityListBlock = styled.div`
-    flex: 1;
-    padding: 20px 32px;
-    padding-bottom: 48px;
-    overflow-y: auto;
-`;
+import NecessityItem from '../NecessityItem/NecessityItem';
+import './NecessityList.css';
 
 interface Props {
   necessities: any;
@@ -18,7 +11,8 @@ function NecessityList(props: Props) {
   const { necessities } = props;
 
   return (
-    <NecessityListBlock>
+    <div className="necessity-list-block">
+
       {necessities.map((necessity: any) => (
         <NecessityItem
           key={necessity.id}
@@ -28,7 +22,8 @@ function NecessityList(props: Props) {
           price={necessity.price}
         />
       ))}
-    </NecessityListBlock>
+
+    </div>
   );
 }
