+++ orange/src/components/Necessity/NecessityTemplate/NecessityTemplate.css
@@ -1,0 +1,17 @@
+.necessity-template-block {
+    width: 512px;
+    height: 600px;
+
+    position: relative;
+
+    background: white;
+    border-radius: 16px;
+    box-shadow: 0 0 8px 0 rgba(0, 0, 0, 0.04);
+
+    margin: 0 auto;
+
+    margin-top: 80px;
+    margin-bottom: 32px;
+    display: flex;
+    flex-direction: column;
+}
